+++ universal_test_runner/matchers.py
@@ -118,12 +118,14 @@ elixir = Matcher.basic_builder("elixir", "mix.exs", "mix test")
 rust = Matcher.basic_builder("rust", "Cargo.toml", "cargo test")
 clojure = Matcher.basic_builder("clojure", "project.clj", "lein test")
 exercism = Matcher.basic_builder("exercism", ".exercism", "exercism test --")
+advent_of_code = Matcher.basic_builder("advent of code", "advent", "./advent")
 
 # these are checked in order
 ALL_MATCHERS: list[Matcher] = [
     justfile,
     exercism,
     makefile,
+    advent_of_code,
     django,
     # anything that could run pytest should go before it
     pytest,